Improve code coverage
- Add script to extract PHPDoc type tests from phpstan/phpdoc-parser and
  use it to update `PhpDocTest`
- Add closure template support to `Regex::PHPDOC_TYPE`
- Add `Regex::INVISIBLE_CHAR`
- In `Get::code()`:
  - Add `$constants` parameter that maps substrings to constant
    identifiers
  - Do not escape CR or LF in multiline mode
  - Do not escape UTF-8 leading or continuation bytes
  - Always escape control characters
  - Escape blank/ignorable characters
  - Remove unnecessary backslashes
  - In arrays with string and integer keys, suppress numeric keys if
    they are numbered consecutively from `0`
- In `PhpDoc`, recognise `@template` syntax `as <type>` in addition to
  `of <type>`

+++ src/Toolkit/Core/Utility/Str.php
@@ -617,7 +617,7 @@ final class Str extends AbstractUtility
         bool $normalise = true
     ): float {
         if ($string1 === '' && $string2 === '') {
-            return (float) 0;
+            return 0.0;
         }
 
         if ($normalise) {
@@ -645,7 +645,7 @@ final class Str extends AbstractUtility
         bool $normalise = true
     ): float {
         if ($string1 === '' && $string2 === '') {
-            return (float) 1;
+            return 1.0;
         }
 
         if ($normalise) {
@@ -707,7 +707,7 @@ final class Str extends AbstractUtility
         int $size
     ): float {
         if (strlen($string1) < $size && strlen($string2) < $size) {
-            return (float) 1;
+            return 1.0;
         }
 
         if ($normalise) {
